CheckPolkitPrivs: implement new check for files put into rules.d dirs
Until now we didn't restrict files installed in /etc/polkit-1/rules.d
and /usr/share/polkit-1/rules.d. Files installed here can circument our
polkit-default-privs, however, and need to be carefully reviewed.

A new subpackage in polkit-default-privs named polkit-whitelisting
contains the whitelistings performed by the security team in JSON
format. This whitelistings file is embedded into rpmlint-mini and
configured via rpmlint config option "PolkitRulesWhitelist".

For being able to restrict whitelistings even to certain packages we are
now also checking for the correct package name. This prevents unrelated
packages to hijack the whitelisting that was intended for another
package.

The content of the rules files is also whitelisted via a hexdigest. Each
whitelisting entry carries a digest of the rules file that has been
checked. This should be the default. For special cases the digest
checking can also be disabled. This is for example the case for the
dynamically generated polkit-default-privs.