Whitelisting: avoid duplicate checks / error messages for the same files
Since we're potentially checking groups of related files for correct
digests, we'll end up performing the same checks over again and again,
since we iterate over each individual file in the RPM and check for
whitelisting entries for it.

This already causes duplicated error messages in some cases and is also
inefficient in case of matching Whitelisting entries. Therefore keep
track of paths we already positively or negatively verified and skip
them when they occur again.